requirements.txt
crispy-bootstrap5
dateparser
debugpy
Django
django-classy-tags
django-cms
django-colorfield
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-extensions
django-filer
django-mathfilters
django-mptt
django-oauth-toolkit
django-organizations
django-registration
django-sekizai
django-tempus-dominus
django-treebeard
django-webpack-loader
djangocms-admin-style
djangocms-bootstrap4
djangocms-file
djangocms-icon
djangocms-link
djangocms-picture
djangocms-style
djangocms-text-ckeditor
djangocms-video
djangorestframework
easy_thumbnails
html5lib
Pillow
psycopg2-binary
pytz
six